#  Fix typo in TryReadUntill and optimize it and JsonReader for the common case