--- a/src/System.Text.JsonLab/System/Text/Json/JsonConstants.cs
+++ b/src/System.Text.JsonLab/System/Text/Json/JsonConstants.cs
@@ -25,6 +25,7 @@ namespace System.Text.JsonLab
         public const byte Space = (byte)' ';
         public const byte CarriageReturn = (byte)'\r';
         public const byte LineFeed = (byte)'\n';
+        public const byte Tab = (byte)'\t';
         public const byte ListSeperator = (byte)',';
         public const byte KeyValueSeperator = (byte)':';
         public const byte Quote = (byte)'"';
@@ -36,6 +37,10 @@ namespace System.Text.JsonLab
         public static readonly byte[] TrueValue = { (byte)'t', (byte)'r', (byte)'u', (byte)'e' };
         public static readonly byte[] FalseValue = { (byte)'f', (byte)'a', (byte)'l', (byte)'s', (byte)'e' };
         public static readonly byte[] NullValue = { (byte)'n', (byte)'u', (byte)'l', (byte)'l' };
+
+        public static readonly byte[] Delimiters = { ListSeperator, CloseBrace, CloseBracket, CarriageReturn, LineFeed, Space, Tab };
+
+        public static readonly byte[] WhiteSpace = { Space, CarriageReturn, LineFeed, Tab };
 
         #endregion Common values
     }