- Fixed potential integer overflow in BufferReader
- Also added initial implementations of some APIs from https://github.com/dotnet/corefxlab/issues/2408.

``` ini

BenchmarkDotNet=v0.10.14.683-nightly, OS=Windows 10.0.17713
Intel Core i7-6700 CPU 3.40GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=2.2.100-preview1-009015
  [Host]     : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT


```
|                                            Method |   TestCase |            Mean |         Error |        StdDev |          Median | Scaled | ScaledSD |  Gen 0 | Allocated |
|-------------------------------------------------- |----------- |----------------:|--------------:|--------------:|----------------:|-------:|---------:|-------:|----------:|
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |  **BasicJson** |       **467.02 ns** |      **2.406 ns** |      **2.250 ns** |       **467.22 ns** |   **1.00** |     **0.00** |      **-** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |  BasicJson |     1,552.24 ns |     91.529 ns |     85.617 ns |     1,517.03 ns |   3.32 |     0.18 | 0.0076 |      40 B |
|                                                   |            |                 |               |               |                 |        |          |        |           |
|              **ReaderSystemTextJsonLabSpanEmptyLoop** | **HelloWorld** |        **77.57 ns** |      **1.566 ns** |      **3.055 ns** |        **76.09 ns** |   **1.00** |     **0.00** |      **-** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop | HelloWorld |       436.67 ns |      8.546 ns |     15.190 ns |       429.40 ns |   5.64 |     0.28 | 0.0091 |      40 B |
|                                                   |            |                 |               |               |                 |        |          |        |           |
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |   **Json400B** |       **731.84 ns** |     **14.576 ns** |     **20.905 ns** |       **726.56 ns** |   **1.00** |     **0.00** |      **-** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |   Json400B |     2,035.42 ns |     25.197 ns |     19.672 ns |     2,032.81 ns |   2.78 |     0.08 | 0.0114 |      56 B |
|                                                   |            |                 |               |               |                 |        |          |        |           |
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |  **Json400KB** |   **668,692.86 ns** |  **5,753.710 ns** |  **5,100.513 ns** |   **669,685.16 ns** |   **1.00** |     **0.00** |      **-** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |  Json400KB | 1,612,234.29 ns | 32,188.072 ns | 58,041.734 ns | 1,614,183.79 ns |   2.41 |     0.09 |      - |       0 B |
|                                                   |            |                 |               |               |                 |        |          |        |           |
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |    **Json4KB** |     **6,610.56 ns** |    **125.760 ns** |    **117.636 ns** |     **6,585.90 ns** |   **1.00** |     **0.00** |      **-** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |    Json4KB |    15,389.81 ns |    250.302 ns |    221.886 ns |    15,332.40 ns |   2.33 |     0.05 | 0.0916 |     416 B |

This change improves perf for the multi-segment case by almost 2x, compared to the previous results: https://github.com/dotnet/corefxlab/pull/2407#discussion_r206368971

We are still ~3x slower than the single-segment case.

cc @JeremyKuhne, @davidfowl, @pakrym