@ahsonkhan 2018-08-01
Alternatively, we could allocate and copy the data from multiple segments into a single segment before parsing it. In micro-benchmarks, this approach performs well, but there are likely performance implications of allocating in memory-constrained environments.

``` ini

BenchmarkDotNet=v0.10.14.683-nightly, OS=Windows 10.0.17713
Intel Core i7-6700 CPU 3.40GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=2.2.100-preview1-009015
  [Host]     : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT


```
|                                            Method |   TestCase |          Mean |         Error |        StdDev |        Median | Scaled | ScaledSD |    Gen 0 |    Gen 1 |    Gen 2 | Allocated |
|-------------------------------------------------- |----------- |--------------:|--------------:|--------------:|--------------:|-------:|---------:|---------:|---------:|---------:|----------:|
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |  **BasicJson** |     **478.40 ns** |      **9.585 ns** |     **14.049 ns** |     **476.05 ns** |   **1.00** |     **0.00** |        **-** |        **-** |        **-** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |  BasicJson |     652.55 ns |     12.707 ns |     13.597 ns |     649.00 ns |   1.37 |     0.05 |   0.0639 |        - |        - |     272 B |
|                                                   |            |               |               |               |               |        |          |          |          |          |           |
|              **ReaderSystemTextJsonLabSpanEmptyLoop** | **HelloWorld** |      **86.16 ns** |      **2.486 ns** |      **7.331 ns** |      **84.77 ns** |   **1.00** |     **0.00** |        **-** |        **-** |        **-** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop | HelloWorld |     208.25 ns |      4.105 ns |      6.629 ns |     206.74 ns |   2.43 |     0.21 |   0.0131 |        - |        - |      56 B |
|                                                   |            |               |               |               |               |        |          |          |          |          |           |
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |   **Json400B** |     **771.06 ns** |     **15.338 ns** |     **43.759 ns** |     **755.96 ns** |   **1.00** |     **0.00** |        **-** |        **-** |        **-** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |   Json400B |     972.28 ns |     18.875 ns |     24.543 ns |     969.06 ns |   1.26 |     0.08 |   0.1087 |        - |        - |     464 B |
|                                                   |            |               |               |               |               |        |          |          |          |          |           |
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |  **Json400KB** | **691,002.81 ns** | **11,195.132 ns** |  **9,348.442 ns** | **691,303.02 ns** |   **1.00** |     **0.00** |        **-** |        **-** |        **-** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |  Json400KB | 833,838.41 ns | 16,337.023 ns | 22,362.301 ns | 833,952.95 ns |   1.21 |     0.04 | 110.3516 | 110.3516 | 110.3516 |  428000 B |
|                                                   |            |               |               |               |               |        |          |          |          |          |           |
|              **ReaderSystemTextJsonLabSpanEmptyLoop** |    **Json4KB** |   **6,957.58 ns** |    **139.089 ns** |    **387.724 ns** |   **6,832.62 ns** |   **1.00** |     **0.00** |        **-** |        **-** |        **-** |       **0 B** |
| ReaderSystemTextJsonLabMultiSpanSequenceEmptyLoop |    Json4KB |   7,860.30 ns |    177.002 ns |    210.709 ns |   7,841.40 ns |   1.13 |     0.07 |   1.0529 |        - |        - |    4440 B |